added requested PR changes, broken out files
